% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Comp_ARI}
\alias{Comp_ARI}
\title{Assess clustering performace}
\usage{
Comp_ARI(rsm_list, clu, ref_label)
}
\arguments{
\item{rsm_list}{the reconstruction dissimilarity matrix list}

\item{clu}{the potential cluster number which could be obtained from hierarchical clustering heatmap of Get_hclust}

\item{ref_label}{the available or known cell type label}
}
\value{
the clustering performance
}
\description{
Calculate ARI to evaluate clustering performance in combination with known cell types
}
\examples{
true_label <- load('syth_label.RData')
k_clu <- 6
ARI_test <- Comp_ARI(rsm,k_clu,true_label)
}
