% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Get_Cosine}
\alias{Get_Cosine}
\title{Process function}
\usage{
Get_Cosine(j, data)
}
\arguments{
\item{list}{}
}
\value{
Cosine dissimilarity object
}
\description{
Get the corresponding dissimilarity matrix using single-cell methylation data
}
