% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Get_Dual}
\alias{Get_Dual}
\title{Process function}
\usage{
Get_Dual(j, data)
}
\arguments{
\item{list}{}
}
\value{
Dual dissimilarity object
}
\description{
Get the corresponding dissimilarity matrix using single-cell methylation data
}
