% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Get_hclust}
\alias{Get_hclust}
\title{Hierarchical clustering visualization}
\usage{
Get_hclust(rcdm)
}
\arguments{
\item{rcdm}{the reconstructed dissimilarity matrix}
}
\value{
hierarchical clustering result
}
\description{
Input the reconstructed dissimilarity matrix and output the hierarchical clustering result to help judge the number of potential cell clusters.
}
\examples{
Get_hclust(rsm)

}
