% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Get_Pearson}
\alias{Get_Pearson}
\title{Process function}
\usage{
Get_Pearson(j, data)
}
\arguments{
\item{list}{}
}
\value{
Pearson dissimilarity object
}
\description{
Get the corresponding dissimilarity matrix using single-cell methylation data
}
