% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Get_resm}
\alias{Get_resm}
\title{Dissimilarity matrix reconstruction}
\usage{
Get_resm(d1, d2, d3, nei_k, is_scale)
}
\arguments{
\item{d1}{one of the three dissimilarity matrices, including Pearson_dism, Cosine_dism and Dual_dism}

\item{d2}{one of the three dissimilarity matrices, including Pearson_dism, Cosine_dism and Dual_dism}

\item{d3}{one of the three dissimilarity matrices, including Pearson_dism, Cosine_dism and Dual_dism}

\item{nei_k}{the n_neighbors to be used in UMAP transformation. The recommended value range is 5~30.}

\item{is_scale}{whether to use normalization for reconstructing dissimilarity matrix.}
}
\value{
List which is composed of normalized reconstructed dissimilarity matrix,and the other three expanded dissimilarity matrices
}
\description{
The input distance matrices are embedded and fused, and the normalized reconstructed distance matrix is output
}
\examples{
dism_pearson <- Output_DISM(k_cpu=8,data_cell,method='Pearson')
dism_cosine  <- Output_DISM(k_cpu=8,data_cell,method='Cosine')
dism_dual    <- Output_DISM(k_cpu=8,data_cell,method='Dual')
rsm          <- Get_resm(dism_cosine,dism_dual,dism_pearson,6,is_scale="TRUE")

}
