% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{nm}
\alias{nm}
\title{Normalization}
\usage{
nm(x)
}
\arguments{
\item{x}{vectors or matrix}

\item{method}{the normalization method to be used. This must be one of "Norm" ,or "scale"}
}
\value{
vectors or matrix
}
\description{
Normalize a vector or matrix
}
\examples{
x_nm <- nm(x,method='Norm')

}
