% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Output_DISM}
\alias{Output_DISM}
\title{Dissimilarity object}
\usage{
Output_DISM(k_cpu, data, method)
}
\arguments{
\item{k_cpu}{number of CPU cores for parallel computing}

\item{data}{R list which is composed of dataframes of cell methylation information, including "Chromosome number", "locus position", and "methylation_level(0 or 1)"}

\item{method}{the distance measure to be used. This must be one of "Cosine" ,"Dual" or "Pearson"}
}
\value{
the corresponding dissimilarity matrix
}
\description{
Get the corresponding dissimilarity matrix using single-cell methylation data
}
\examples{
load('data_bd.RData')
view(data_BLOOD)
dism_pearson <- Output_DISM(k_cpu=8,data_BLOOD,method='Pearson')

}
