% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{Comp_ARI}
\alias{Comp_ARI}
\title{Assess clustering performace}
\usage{
Comp_ARI(hc_obj, clu, ref_label)
}
\arguments{
\item{hc_obj}{the hierarchical clustering object from reconstructed dissimilarity matrix}

\item{clu}{the potential cluster number which could be obtained from hierarchical clustering heatmap of Get_hclust}

\item{ref_label}{the available or known cell type label}
}
\value{
the clustering performance
}
\description{
Calculate ARI to evaluate clustering performance
}
\examples{
true_label <- load('ref_label.RData')
k_clu <- 6
ARI_test <- Comp_ARI(hc_obj,k_clu,true_label)
}
