% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINCEF_source.R
\name{mat_F}
\alias{mat_F}
\title{process function}
\usage{
mat_F(df)
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
Obtain common methylation sites for any two cells
}
\examples{
df_test <- mat_F(df)

}
